+++ IPassPlusSdk/src/main/java/com/sdk/ipassplussdk/model/response/facesimilarity/Data.kt
@@ -2,7 +2,7 @@ package com.sdk.ipassplussdk.model.response.facesimilarity
 
 data class Data(
     val compareResult: CompareResult,
-    val facePercentage: Int,
+    val facePercentage: Double,
     val status: String,
     val success: Boolean
 )
